# openapi-appengine.yaml
swagger: "2.0"
...
info:
  description: "Dantion Endpoints API."
  title: "Endpoints Dantion"
  version: "1.0.0"
host: "dantion.et.r.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
  - "http"

# components:
#   securitySchemes:
#     bearerAuth:
#       type: http
#       scheme: bearer
#       bearerFormat: JWT
#   responses:
#     UnauthorizedError:
#       description: Access token is missing or invalid

# This section requires all requests to any path to require an API key.
security:
  # - api_key: ['AIzaSyAMiyLWcIlzAZnQWciTMLQjIARqla9zzxc']
  - api_key: []
securityDefinitions:
  # This section configures basic authentication with an API key.
  api_key:
    type: "apiKey"
    name: "key"
    in: "query"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
  # This section configures authentication using Google API Service Accounts
  # to sign a json web token. This is mostly used for server-to-server
  # communication.
  # google_jwt:
  #   authorizationUrl: ""
  #   flow: "implicit"
  #   type: "oauth2"
  #   # This must match the 'iss' field in the JWT.
  #   x-google-issuer: "jwt-client.endpoints.sample.google.com"
  #   # Update this with your service account's email address.
  #   x-google-jwks_uri: "https://www.googleapis.com/service_accounts/v1/jwk/YOUR-SERVICE-ACCOUNT-EMAIL"
  # # This section configures authentication using Google OAuth2 ID Tokens.
  # # ID Tokens can be obtained using OAuth2 clients, and can be used to access
  # # your API on behalf of a particular user.
  # google_id_token:
  #   authorizationUrl: ""
  #   flow: "implicit"
  #   type: "oauth2"
  #   x-google-issuer: "accounts.google.com"
  #   x-google-jwks_uri: "https://www.googleapis.com/oauth2/v1/certs"
paths:
  "/admins/userRole":
    patch:
      tags:
        - "Admin"
      security:
        - Bearer: []
      description: "Update User Role"
      operationId: "UserRole"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: Admin Update User Role
          description: Request body for update user role.
          schema:
            $ref: "#/definitions/AdminUpdateUserRoleBody"
      responses:
        201:
          description: "Success update user role"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
  "/detections/{id}":
    get:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Get Detection Using ID"
      operationId: "detectionId"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success get detail detection"
          schema:
            $ref: "#/definitions/DetectionSuccessResponse"
        400:
          description: "failed get detail detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
    delete:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Delete Detection ID"
      operationId: "deleteDetection"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
        400:
          description: "failed delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
definitions:
  StatusMessageResponse:
    type: "object"
    properties:
      status:
        type: "string"
      message:
        type: "string"
...
  AdminUpdateUserRoleBody:
    type: "object"
    properties:
      userId:
        type: "string"
      role:
        type: "string"
        enum:
          - "polisi"
          - "ambulan"
          - "damkar"
...
  Detection:
    type: "object"
    properties:
      id:
        type: "string"
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      recordUrl:
        type: "string"
        format: "binary"
      type:
        type: "string"
        enum:
          - "kejahatan"
          - "kecelakaan"
          - "kebakaran"
      isValid:
        type: "boolean"
        default: false
      userId:
        type: "string"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
  DetectionSuccessResponse:
    type: "object"
    properties:
      status:
        type: "string"
      message:
        type: "string"
      user:
        $ref: "#/definitions/Detection"
...